internationalisation, IE support, markers
The internationalisation is done with babel and jinja2.
To generate the index files (en, fr and de), running generate.py is
enough. To generate the .pot, .po and .mo file, babel needs to be
installed. The batch scripts can help for the parameters.
IE support for the CSV export has been achieved by changing the way the
file is exported.
The markers are now displayed with a letter.